+++ src/main/java/online/kingdomkeys/kingdomkeys/lib/Strings.java
@@ -806,6 +806,7 @@ public class Strings {
         wayfinder = "wayfinder",
         gummiPhone = "gummiphone",
 		gummiShipBlueprint = "gummi_ship_blueprint",
+		gummiShipBlueprintCreative = "gummi_ship_blueprint_creative",
         
         powerBoost = "powerboost",
         magicBoost = "magicboost",